Fix quiz end screen
* Fix quiz end screen

Quiz was routing to initial route because it was designed to sync
customer state this way.

This was causing undesired behaviour such as requiring customer
login instead of going to safe area when quiz suggested so.

Using quiz end screen will leave this responsibility to chat AI
decision.

* Add page query param on mainboard

With this we can route customer to a specific mainboard tab.

* Enable navigation with routes pattern

It will enable the usage of query params so we can navigate to main
board in a specific tab.

* Check app state before routing to quiz end screen

Check app state is required so app state can reload app modules
with a fresh /me request.